インベントリーの途中経過

+++ Game/IntVector3.h
@@ -4,7 +4,7 @@ struct IntVector3{
 	int x, y, z;
 
 	constexpr IntVector3( int x, int y, int z ) : x( x ), y( y ), z( z ){}
-	explicit IntVector3(const CVector3& v) : x(std::floor(v.x)), y( std::floor( v.y ) ), z( std::floor( v.z ) ){}
+	explicit IntVector3(const CVector3& v) : x(int(std::floor(v.x))), y( int(std::floor( v.y )) ), z( int(std::floor( v.z )) ){}
 
 	bool operator==( const IntVector3& rhs ) const{
 		return rhs.x == x && rhs.y == y && rhs.z == z;
